% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{bracketed_tex}
\alias{bracketed_tex}
\title{Bracket controlled tex output}
\usage{
bracketed_tex(x, brackets = list(c("{", "}"), c("[", "]"), c("(", ")")))
}
\arguments{
\item{x}{yac_symbol}

\item{brackets}{Symbols to use as brackets}
}
\description{
Make brackets to different types.
}
