% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{btex}
\alias{btex}
\title{Bracket controlled tex output}
\usage{
btex(x, brackets = list(c("{", "}"), c("[", "]"), c("(", ")")))
}
\arguments{
\item{x}{yac_symbol}

\item{brackets}{Symbols to use as brackets. Reused as necesary.}
}
\description{
Make brackets to different types.
}
