% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{hbtex}
\alias{hbtex}
\title{Make hard brackets}
\usage{
hbtex(x)
}
\arguments{
\item{x}{A tex string}
}
\description{
Make hard brackets
}
\examples{
x <- paste0("\\\\left( \\\\left( 2 \\\\sin \\\\left( x + 2\\\\right) \\\\right)  ^{2} + ",
            "\\\\left( a + b\\\\right)  ^{2}\\\\right)  ^{3}")
btex(x)
hbtex(x)

if (requireNamespace("Ryacas", quiet = TRUE)) {
  A <- Ryacas::ysym(matrix(1:9, 3, 3))
  x <- Ryacas::tex(A)
  x
  btex(x)
  hbtex(x)
}
}
