% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit-product.R
\name{make_products_explicit}
\alias{make_products_explicit}
\title{Make implicit products explicit}
\usage{
make_products_explicit(
  x,
  allowed_functions = getGroupMembers("Math"),
  allowed_constants = c("pi")
)
}
\arguments{
\item{x}{user input}

\item{allowed_functions}{functions (sequence of characters) that should not have products, e.g. \code{cos}, \code{sin}}
}
\description{
Make implicit products explicit
}
\examples{
make_products_explicit("2x")
make_products_explicit("2(1+2)")
make_products_explicit("x(y)")
make_products_explicit("(1+2)3") # Not allowed by definition
make_products_explicit("(1+2)x") # Is allowed
make_products_explicit("x(2+1) + sin(y)")
make_products_explicit("sin(2+1) + x(y+1)")
make_products_explicit("(2x)(y+1)")

}
