% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe-eval.R
\name{prepare_input}
\alias{prepare_input}
\title{Prepare input}
\usage{
prepare_input(
  x,
  replace_comma = TRUE,
  insert_products = TRUE,
  allowed_functions = getGroupMembers("Math"),
  allowed_constants = c("pi")
)
}
\description{
Prepare input
}
