% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe-eval.R
\name{safe_eval}
\alias{safe_eval}
\title{Safely eval a string as an expression}
\usage{
safe_eval(
  x,
  vars = NULL,
  replace_comma = TRUE,
  insert_products = TRUE,
  allowed_functions = getGroupMembers("Math"),
  allowed_constants = c("pi")
)
}
\arguments{
\item{x}{string to evaluate}

\item{vars}{list with variable values}

\item{replace_comma}{Replace ',' with '.' before parsing}

\item{insert_products}{Make implicit products explicit}

\item{allowed_functions}{functions (sequence of characters) that should not have products, e.g. \code{cos}, \code{sin}}
}
\description{
Safely eval a string as an expression
}
\examples{
safe_eval("2+2")

}
