% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{btex}
\alias{btex}
\title{Bracket controlled tex output}
\usage{
btex(x, brackets = list(c("\\\\{", "\\\\}"), c("[", "]"), c("(", ")")))
}
\arguments{
\item{x}{A tex string}

\item{brackets}{Symbols to use as brackets. Reused as necesary.}
}
\description{
Make brackets to different types.
}
\examples{
x <- paste0("\\\\left( \\\\left( 2 \\\\sin \\\\left( x + 2\\\\right) \\\\right)  ^{2} + ",
            "\\\\left( a + b\\\\right)  ^{2}\\\\right)  ^{3}")
btex(x)
hbtex(x)
}
