% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{btex}
\alias{btex}
\title{Bracket controlled tex output}
\usage{
btex(x, brackets = list(c("\\\\{", "\\\\}"), c("[", "]"), c("(", ")")))
}
\arguments{
\item{x}{yac_symbol}

\item{brackets}{Symbols to use as brackets. Reused as necesary.}
}
\description{
Make brackets to different types.
}
\examples{
e <- Ryacas::ysym("((2*Sin(2+x))^2 + (a+b)^2)^3")
Ryacas::tex(e)
btex(e)
hbtex(e)
}
