% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-reply-answer.R
\name{compare_reply_answer}
\alias{compare_reply_answer}
\title{Compare a reply to an answer}
\usage{
compare_reply_answer(
  reply,
  answer,
  compare_grid = NULL,
  replace_comma = TRUE,
  ans_tol = 0.01
)
}
\arguments{
\item{reply}{user-provided reply as string}

\item{answer}{the true answer provided as \code{yac_symbol}}

\item{compare_grid}{grid to perform comparisons on; must have columns}

\item{replace_comma}{Replace ',' with '.' before parsing}

\item{ans_tol}{tolerance for comparisons}
}
\description{
The comparison is done via a numerical grid.
}
