% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-reply-answer.R
\name{compare_reply_answer}
\alias{compare_reply_answer}
\title{Compare a reply to an answer}
\usage{
compare_reply_answer(
  reply,
  answer,
  compare_grid = NULL,
  replace_comma = TRUE,
  ans_tol = 0.01,
  insert_products = TRUE,
  allowed_functions = getGroupMembers("Math")
)
}
\arguments{
\item{reply}{user-provided reply as string/character matrix (will be evaluated safely)}

\item{answer}{the true answer provided as an \code{R} expression (or list of list of expressions; first level are columns, then rows)}

\item{compare_grid}{grid to perform comparisons on; must have columns}

\item{replace_comma}{Replace ',' with '.' before parsing}

\item{ans_tol}{tolerance for comparisons}

\item{insert_products}{Make implicit products explicit}

\item{allowed_functions}{functions (sequence of characters) that should not have products, e.g. \code{cos}, \code{sin}}
}
\description{
The comparison is done via a numerical grid.
}
\examples{
compare_reply_answer(reply = "2yx^2", 
                     answer = expression("2*y*x^2"), 
                     compare_grid = expand.grid(x = seq(-10, 10, len = 10), 
                                                y = seq(-10, 10, len = 10)))
compare_reply_answer(reply = "2.1yx^2", 
                     answer = expression("2*y*x^2"), 
                     compare_grid = expand.grid(x = seq(-10, 10, len = 10), 
                                                y = seq(-10, 10, len = 10)))
compare_reply_answer(reply = "2yx^2+0.01", 
                     answer = expression("2*y*x^2"), 
                     compare_grid = expand.grid(x = seq(-10, 10, len = 10), 
                                                y = seq(-10, 10, len = 10)), 
                     ans_tol = 1)
                     
rep <- matrix(c("(1+1)*x", "3*x"), nrow = 1, ncol = 2)
ans <- list(list(expression(2*x)), list(expression(3*x)))
compare_reply_answer(reply = rep,
                     answer = ans,
                     compare_grid = expand.grid(x = seq(-10, 10, len = 10)),
                     ans_tol = 1)
}
