% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeparse.R
\name{safe_eval}
\alias{safe_eval}
\title{Safely eval a string as an expression}
\usage{
safe_eval(x, vars = NULL, replace_comma = TRUE)
}
\arguments{
\item{x}{string to evaluate}

\item{vars}{list with variable values}

\item{replace_comma}{Replace ',' with '.' before parsing}
}
\description{
Safely eval a string as an expression
}
\examples{
safe_eval("2+2")

}
